+++ apps/web/core/components/editor/rich-text/description-input/root.tsx
@@ -275,7 +275,7 @@ export const DescriptionInput = observer(function DescriptionInput(props: Props)
               return asset_id;
             } catch (error) {
               console.log("Error in uploading asset:", error);
-              throw new Error("Asset upload failed. Please try again later.");
+              throw new Error(t("issue.comments.upload.error"));
             }
           }}
           duplicateFile={async (assetId: string) => {
@@ -288,7 +288,7 @@ export const DescriptionInput = observer(function DescriptionInput(props: Props)
               });
               return asset_id;
             } catch {
-              throw new Error("Asset duplication failed. Please try again later.");
+              throw new Error(t("issue.comments.duplicate.error"));
             }
           }}
         />